grok/requirements.txt
# Installed with NuPIC
dbutils==1.1
numpy==1.9.2
mysql-python==1.2.5
pytest==2.4.2
pytest-cov==1.6
pytest-xdist==1.8
unittest2==0.5.1
validictory==0.9.1

# Required by YOMP
boto==2.38.0
fabric==1.9.0
msgpack-python==0.3.0
paste==1.7.5.1
pytz==2015.2
requests==2.0.1
sphinx==1.1.3
supervisor==3.1.3
uwsgi==2.0.4
web.py==0.37
sqlalchemy==0.9.4
alembic==0.6.7

# Required by helper scripts and pipeline
PyYAML==3.10

# Required by tests
mock==1.0.1

# Other numenta projects
nta.utils>=0.0.0
htmengine>=0.0.0
